Remove old proof of status component and rename new one
