+++ src/applications/personalization/profile/components/proof-of-veteran-status/ProofOfVeteranStatus.unit.spec.jsx
@@ -1,16 +1,29 @@
 import React from 'react';
 import { expect } from 'chai';
-import { fireEvent } from '@testing-library/react';
+import * as api from '~/platform/utilities/api';
+import { waitFor } from '@testing-library/react';
 import sinon from 'sinon';
-
-import * as generatePdfModule from '~/platform/pdf';
 import { renderWithProfileReducers } from '../../tests/unit-test-helpers';
 import ProofOfVeteranStatus from './ProofOfVeteranStatus';
 
-const eligibleServiceHistoryItem = {
+const serviceHistoryItemOlder = {
   branchOfService: 'Air Force',
-  beginDate: '2009-04-12',
-  endDate: '2013-04-11',
+  beginDate: '2000-01-01',
+  endDate: '2001-01-01',
+  personnelCategoryTypeCode: 'V',
+  characterOfDischargeCode: 'A',
+};
+const serviceHistoryItemMiddle = {
+  branchOfService: 'Air Force',
+  beginDate: '2010-01-01',
+  endDate: '2011-01-01',
+  personnelCategoryTypeCode: 'V',
+  characterOfDischargeCode: 'A',
+};
+const serviceHistoryItemNewer = {
+  branchOfService: 'Air Force',
+  beginDate: '2020-01-01',
+  endDate: '2021-01-01',
   personnelCategoryTypeCode: 'V',
   characterOfDischargeCode: 'A',
 };
@@ -47,8 +60,16 @@ const nonEligibility = {
   ],
 };
 
-function createBasicInitialState(serviceHistory, eligibility) {
+function createBasicInitialState(
+  serviceHistory,
+  eligibility,
+  enableToggle = false,
+) {
   return {
+    featureToggles: {
+      // eslint-disable-next-line camelcase
+      veteran_status_card_use_lighthouse_frontend: enableToggle,
+    },
     user: {
       profile: {
         veteranStatus: {
@@ -80,9 +101,18 @@ function createBasicInitialState(serviceHistory, eligibility) {
 }
 
 describe('ProofOfVeteranStatus', () => {
+  let apiRequestStub;
+
+  beforeEach(() => {
+    apiRequestStub = sinon.stub(api, 'apiRequest');
+  });
+  afterEach(() => {
+    apiRequestStub.restore();
+  });
+
   describe('when it exists', () => {
     const initialState = createBasicInitialState(
-      [eligibleServiceHistoryItem],
+      [serviceHistoryItemMiddle],
       confirmedEligibility,
     );
 
@@ -90,77 +120,378 @@ describe('ProofOfVeteranStatus', () => {
       const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
         initialState,
       });
-      expect(view.queryByText(/Proof of Veteran status/)).to.exist;
+      const heading = view.getAllByText(/Proof of Veteran status/i);
+      expect(heading).to.have.lengthOf.above(0);
     });
 
-    it('should render description copy', () => {
+    it('should render description copy', async () => {
       const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
         initialState,
       });
-      expect(
-        view.queryByText(
-          /get discounts offered to Veterans at many restaurants/i,
-        ),
-      ).to.exist;
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /This card identifies a Veteran of the U.S. Uniformed Services./i,
+          ),
+        ).to.exist;
+      });
     });
 
-    it('should render mobile app callout', () => {
+    it('should render mobile app callout', async () => {
+      const mockData = {
+        data: {
+          id: '',
+          type: 'veteran_status_confirmations',
+          attributes: { veteranStatus: 'confirmed' },
+        },
+      };
+
+      apiRequestStub.resolves(mockData);
       const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
         initialState,
       });
-      expect(
-        view.queryByText(/Get proof of Veteran Status on your mobile device/i),
-      ).to.exist;
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /Get proof of Veteran Status on your mobile device/i,
+          ),
+        ).to.exist;
+      });
+    });
+  });
+
+  describe('should fetch verification status on render', () => {
+    let initialState = createBasicInitialState(
+      [serviceHistoryItemMiddle],
+      confirmedEligibility,
+      true,
+    );
+
+    it('displays the card successfully', async () => {
+      const mockData = {
+        data: {
+          id: '',
+          type: 'veteran_status_confirmations',
+          attributes: { veteranStatus: 'confirmed' },
+        },
+      };
+
+      apiRequestStub.resolves(mockData);
+
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      sinon.assert.calledWith(
+        apiRequestStub,
+        '/profile/vet_verification_status',
+      );
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /Get proof of Veteran status on your mobile device/i,
+          ),
+        ).to.exist;
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with your discharge status records. We can’t provide a Veteran status card for you right now./,
+          ),
+        ).to.not.exist;
+      });
+    });
+
+    it('displays the returned not confirmed message', async () => {
+      const mockData = {
+        data: {
+          id: '',
+          type: 'veteran_status_confirmations',
+          attributes: {
+            veteranStatus: 'not confirmed',
+            notConfirmedReason: 'PERSON_NOT_FOUND',
+          },
+          message: problematicEligibility.message,
+        },
+      };
+
+      apiRequestStub.resolves(mockData);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /Get proof of Veteran Status on your mobile device/i,
+          ),
+        ).to.not.exist;
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with your discharge status records. We can’t provide a Veteran status card for you right now./,
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('handles empty API response', async () => {
+      const mockData = {
+        data: {},
+      };
+      apiRequestStub.resolves(mockData);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            'We’re sorry. There’s a problem with our system. We can’t show your Veteran status card right now. Try again later.',
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('handles API error', async () => {
+      apiRequestStub.rejects(new Error('API Error'));
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      await waitFor(() => {
+        expect(
+          view.getByText(
+            'We’re sorry. There’s a problem with our system. We can’t show your Veteran status card right now. Try again later.',
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('handles a 504 API error with service history', async () => {
+      const mockData = {
+        errors: [
+          {
+            title: 'Gateway Timeout',
+            detail: 'Did not receive a timely response.',
+            code: '504',
+            status: '504',
+          },
+        ],
+      };
+      apiRequestStub.rejects(mockData);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            'We’re sorry. There’s a problem with our system. We can’t show your Veteran status card right now. Try again later.',
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('handles a 403 API error with service history', async () => {
+      const mockData = {
+        errors: [
+          {
+            title: 'Forbidden',
+            detail: 'User does not have access to the requested resource',
+            code: '403',
+            status: '403',
+          },
+        ],
+      };
+      apiRequestStub.rejects(mockData);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            'We’re sorry. There’s a problem with our system. We can’t show your Veteran status card right now. Try again later.',
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('displays not confirmed message if confirmed with no service history', async () => {
+      const mockData = {
+        data: {
+          id: '',
+          type: 'veteran_status_confirmations',
+          attributes: { veteranStatus: 'confirmed' },
+        },
+      };
+      apiRequestStub.resolves(mockData);
+
+      initialState = createBasicInitialState([], problematicEligibility, true);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      sinon.assert.calledWith(
+        apiRequestStub,
+        '/profile/vet_verification_status',
+      );
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with your discharge status records. We can’t provide a Veteran status card for you right now./,
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('displays not confirmed message if not confirmed and no service history', async () => {
+      const mockData = {
+        data: {
+          id: '',
+          type: 'veteran_status_confirmations',
+          attributes: {
+            veteranStatus: 'not confirmed',
+            notConfirmedReason: 'PERSON_NOT_FOUND',
+          },
+          message: problematicEligibility.message,
+        },
+      };
+      apiRequestStub.resolves(mockData);
+      initialState = createBasicInitialState([], problematicEligibility, true);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /Get proof of Veteran Status on your mobile device/i,
+          ),
+        ).to.not.exist;
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with your discharge status records. We can’t provide a Veteran status card for you right now./,
+          ),
+        ).to.exist;
+      });
+    });
+
+    it('displays loading indicator if fetch not complete', async () => {
+      initialState = createBasicInitialState([], problematicEligibility, true);
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      expect(view.getByTestId('proof-of-status-loading-indicator')).to.exist;
     });
   });
 
   describe('when eligible', () => {
     const initialState = createBasicInitialState(
       [
-        eligibleServiceHistoryItem,
+        serviceHistoryItemOlder,
+        serviceHistoryItemMiddle,
+        serviceHistoryItemNewer,
         ineligibleServiceHistoryItem,
         neutralServiceHistoryItem,
       ],
       confirmedEligibility,
     );
 
-    it('should render card if service history contains an eligible discharge despite any other discharges', () => {
+    it('should render card if service history contains an eligible discharge despite any other discharges', async () => {
       const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
         initialState,
       });
 
-      expect(
-        view.queryByAltText(
-          /sample proof of veteran status card featuring name, date of birth, disability rating and period of service/,
-        ),
-      ).to.exist;
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /This status doesn’t entitle you to any VA benefits./,
+          ),
+        ).to.exist;
+      });
     });
 
-    it('should show error message if pdf generation fails', () => {
-      const generatePdfStub = sinon
-        .stub(generatePdfModule, 'generatePdf')
-        .throws(new Error('Some Error'));
+    it('should render the latest service item', async () => {
       const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
         initialState,
       });
-      const link = view.container.querySelector('va-link');
-      const errorMessage =
-        "We're sorry. Something went wrong on our end. Please try to download your Veteran status card later.";
+      await waitFor(() => {
+        expect(view.queryByText(/United States Air Force • 2020–2021/)).to
+          .exist;
+      });
+    });
 
-      expect(link.getAttribute('text')).to.equal(
-        'Download and print your Veteran status card',
-      );
-      expect(view.queryByText(errorMessage)).to.not.exist;
+    it('should render the print button', async () => {
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
 
-      fireEvent.click(link);
+      await waitFor(() => {
+        const link = view.container.querySelector('va-link');
+        expect(link.getAttribute('text')).to.equal(
+          'Print your Proof of Veteran status (PDF)',
+        );
+      });
+    });
+  });
 
-      expect(generatePdfStub.called).to.be.true;
-      expect(view.queryByText(errorMessage)).to.exist;
+  describe('when there is no service history', () => {
+    const initialState = createBasicInitialState([], confirmedEligibility);
+
+    it('should render an error and not the HTML card', async () => {
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      const heading = view.getAllByText(/Proof of Veteran status/);
+      expect(heading).to.have.lengthOf.to.be(1); // only appears once as the section heading, not here as the card heading
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with our system. We can’t show your Veteran status card right now. Try again later./,
+          ),
+        ).to.exist;
+      });
+    });
+  });
+
+  describe('when there is no veteran full name', () => {
+    const initialState = createBasicInitialState(
+      [
+        serviceHistoryItemMiddle,
+        ineligibleServiceHistoryItem,
+        neutralServiceHistoryItem,
+      ],
+      confirmedEligibility,
+    );
+    initialState.vaProfile.hero.userFullName = {
+      first: '',
+      middle: '',
+      last: '',
+      suffix: '',
+    };
+
+    it('should render an error and not the HTML card', async () => {
+      const view = renderWithProfileReducers(<ProofOfVeteranStatus />, {
+        initialState,
+      });
+
+      const heading = view.getAllByText(/Proof of Veteran status/);
+      expect(heading).to.have.lengthOf.to.be(1); // only appears once as the section heading, not here as the card heading
+
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with our system. We can’t show your Veteran status card right now. Try again later./,
+          ),
+        ).to.exist;
+      });
     });
   });
 
   describe('discharge status problem message', () => {
-    it('should render if service history contains neither eligible nor ineligible discharges', () => {
+    it('should render if service history contains neither eligible nor ineligible discharges', async () => {
       const initialState = createBasicInitialState(
         [neutralServiceHistoryItem],
         problematicEligibility,
@@ -169,16 +500,18 @@ describe('ProofOfVeteranStatus', () => {
         initialState,
       });
 
-      expect(
-        view.queryByText(
-          /We’re sorry. There’s a problem with your discharge status records./,
-        ),
-      ).to.exist;
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /We’re sorry. There’s a problem with your discharge status records./,
+          ),
+        ).to.exist;
+      });
     });
   });
 
   describe('ineligibility message', () => {
-    it('should render if service history does not contain an eligible discharge, but does contain an inelible discharge', () => {
+    it('should render if service history does not contain an eligible discharge, but does contain an inelible discharge', async () => {
       const initialState = createBasicInitialState(
         [ineligibleServiceHistoryItem, neutralServiceHistoryItem],
         nonEligibility,
@@ -187,11 +520,13 @@ describe('ProofOfVeteranStatus', () => {
         initialState,
       });
 
-      expect(
-        view.queryByText(
-          /Our records show that you’re not eligible for a Veteran status card./,
-        ),
-      ).to.exist;
+      await waitFor(() => {
+        expect(
+          view.queryByText(
+            /Our records show that you’re not eligible for a Veteran status card./,
+          ),
+        ).to.exist;
+      });
     });
   });
 });
